+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/BlockBody/Internal.hs
@@ -52,6 +52,7 @@ import Cardano.Ledger.Binary (
   encodeNullStrictMaybe,
   serialize',
  )
+import Cardano.Ledger.Block (LeiosEraBlockHeader)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Dijkstra.Era
 import Cardano.Ledger.Dijkstra.Tx (
@@ -113,6 +114,7 @@ type instance MemoHashIndex (DijkstraBlockBodyRaw era) = EraIndependentBlockBody
 
 instance EraBlockBody DijkstraEra where
   type BlockBody DijkstraEra = DijkstraBlockBody DijkstraEra
+  type ProtocolEraBlockHeader h DijkstraEra = LeiosEraBlockHeader h DijkstraEra
   mkBasicBlockBody = mkBasicBlockBodyDijkstra
   txSeqBlockBodyL = lensMemoRawType @DijkstraEra dbbrTxs (\bb p -> bb {dbbrTxs = p})
   hashBlockBody (MkDijkstraBlockBody m) = extractHash $ getMemoBytesHash m
@@ -237,5 +239,8 @@ data PerasKey = PerasKey
 --
 -- NOTE: this function will be replaced with the real implementation from
 -- 'cardano-base' once it's ready.
+--
+-- IMPORTANT: Until Peras is implemented this function must return False, which effectively disables
+-- Peras certificate additon to the block body, hence disables Peras all together
 validatePerasCert :: Nonce -> PerasKey -> PerasCert -> Bool
-validatePerasCert _ _ _ = True
+validatePerasCert _ _ _ = False