+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.24.0.0
 
+* Replace `conwayBbodyTransition` with `bbodyTransition`
 * Change the type of `computeDRepDistr` to take and return a `VotingStakePoolDistr` instead of a `PoolDistr`
 * Rename `reStakePoolDistrL` to `reVotingStakePoolDistrL` and change its type to use `VotingStakePoolDistr`
 * Rename `DRepPulser {dpStakePoolDistr}` to `dpVotingStakePoolDistr` and change the type to `VotingStakePoolDistr`